Move fasta tests into test_phylo_formats
